+++ app/src/main/kotlin/com/wire/android/ui/home/newconversation/contacts/ContactsViewModel.kt
@@ -14,58 +14,27 @@ class ContactsViewModel @Inject constructor(private val navigationManager: Navig
 
     val contactsState by mutableStateOf(
         ContactsState(
-            contacts = listOf(
-                Contact(
-                    id = 1,
-                    "This is first contact"
-                ), Contact(
-                    id = 2,
-                    "test2"
-                ), Contact(
-                    id = 3,
-                    "test3"
-                ),
-                Contact(
-                    id = 4,
-                    "test1"
-                ), Contact(
-                    id = 5,
-                    "test2"
-                ), Contact(
-                    id = 6,
-                    "test3"
-                ),
-                Contact(
-                    id = 7,
-                    "test1"
-                ), Contact(
-                    id = 8,
-                    "test2"
-                ), Contact(
-                    id = 9,
-                    "test3"
-                ),
-                Contact(
-                    id = 10,
-                    "test1"
-                ), Contact(
-                    id = 11,
-                    "test2"
-                ), Contact(
-                    id = 12,
-                    "test3"
-                ),
-                Contact(
-                    id = 13,
-                    "test1"
-                ), Contact(
-                    id = 14,
-                    "test2"
-                ), Contact(
-                    id = 15,
-                    "This is last contact"
+            contacts = buildList {
+                add(
+                    Contact(
+                        id = 1,
+                        "This is first contact"
+                    )
                 )
-            )
+                for (i in 2..11) {
+                    Contact(
+                        id = 1,
+                        "This is $i contact"
+                    )
+                }
+                add(
+                    Contact(
+                        id = 12,
+                        "This is last contact"
+                    )
+                )
+
+            }
         )
     )
