+++ app/src/main/kotlin/com/wire/android/ui/home/newconversation/search/EmptySearchQueryScreen.kt
@@ -1,0 +1,57 @@
+package com.wire.android.ui.home.newconversation.search
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextDecoration
+import com.wire.android.R
+import com.wire.android.ui.common.dimensions
+import com.wire.android.ui.theme.wireColorScheme
+import com.wire.android.ui.theme.wireTypography
+
+
+@Composable
+fun EmptySearchQueryScreen() {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Column(
+            modifier = Modifier
+                .align(Alignment.Center)
+                .fillMaxWidth()
+                .wrapContentHeight()
+                .padding(horizontal = dimensions().spacing48x),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center,
+        ) {
+            Text(
+                text = stringResource(R.string.label_search_people_instruction),
+                style = MaterialTheme.wireTypography.body01.copy(color = MaterialTheme.wireColorScheme.secondaryText),
+                textAlign = TextAlign.Center
+            )
+            Spacer(modifier = Modifier.height(dimensions().spacing16x))
+            Text(
+                text = stringResource(R.string.label_learn_more_searching_user),
+                style = MaterialTheme.wireTypography.body02.copy(
+                    textDecoration = TextDecoration.Underline,
+                    color = MaterialTheme.colorScheme.primary
+                ),
+                modifier = Modifier.clickable {
+                    //TODO: redirect to somewhere ?
+                }
+            )
+        }
+    }
+}
